// src/app/models/empresa.ts
export class EmpresaI {
	id_empresa?: number;
	administrador?: string;
	calificacion?: number;
	email: string;
	email_validado?: boolean;
	foto_empresa?: string;
	giro?: string;
	nombre?: string;
	pagina_web?: string;
	pass: string;
	suscripcion_notificacion?: string;
	telefono?: string;
	ubicacion?: string;
	estado?: string;
}